Recover when reading unknown types from messages
Values in messages are now written by first writing the size of the data
and only then the data. When reading messages this allows to recover in
case the type of the value we're reading is not known by the metatype
system of the client process, by jumping ahead of it and keeping reading
the next values.