ObjectWrapper: Replace LAMBDA_PROP by CUSTOM_PROP
This commit removes the LAMBDA_PROP macro, taking a lambda to create a
property for the object wrapper, and replaces it by the CUSTOM_PROP
macro, which takes an expression instead. The reason is simply, that
I was not able to implement return type deduction for it with lambdas
and C++11.