Objectwrapper: Add support for setters
So far, object wrapper properties only supported reading the value, not
writing it. Now there is a second macro called RW_PROP, which has an
argument 'SetterName'. PROP was renamed to RO_PROP.

For now, setters are called synchronously, support for caching writes
and cross-thread communication for calling setters is not implemented, yet.