WIP: Make ObjectWrapper multi-inheritance ready
Current state is:
It compiles. It has not been verifed, whether something important was
omitted or if the architecture makes sense, in the end. Running the
tests crashes everything. That could be related to this change or be
introduced by an earlier change.

Next steps:
Find out why it crashes. Redesign everything from scratch after it
turned out that the crash is actually a flaw in the design that can't be
fixed.