ObjectWrapper: Make older GCC-versions happy
Older gcc versions seem to not support full template specialization
outside a namespace for templates that were defined inside a namespace
using the syntax `template<> class GammaRay::Mytemplate<Foo> {};`.
So we need to open the namespace, define the specialization there and
close it again.

Also, we used a static-assert without message, which is not supported in
C++11.