ObjectWrapper: Allow expansion of the cache
Consider the following code:
```
class A;
class B : public A { /*...*/ };
B b;
A *a = &b;
auto aHandle = ObjectShadowDataRepository::handleForObject(a);
auto bHandle = ObjectShadowDataRepository::handleForObject(&b);
```
We have an object of type B, but the first time we create a handle for
it, we do so using an A*, thus only wrapping A. Afterwards we try to
create a handle to B for it. Accessing properties of B will crash
because the cache for B doesn't exist.

Now, creating the second handle will expand and update the cache to
include the properties of B.