ObjectWrapper: Add ForeignPointer flag
A pointer property marked as ForeignPointer will be cached in unwrapped
form and only be wrapped on access. This defers the caching of the
pointed-to-object, allowing the developer to control that.

This is useful and necessary e.g. for the QtQuick Item tree, where each
item has a property pointing to the scenegraph-tree. It is though
necessary to create the scenegraph-tree from the root, as else
NonOwningPointer properties would try to access not-yet-cached objects.