ObjectWrapper: CUSTOM_PROP: require explicit type
Previously CUSTOM_PROP auto-deduced the type by putting the expression
into decltype() for the return type of the fetch_##PropertyName method.
This, however, outruled usage of lambdas in CUSTOM_PROP expressions.
Thus it now has the type as a macro argument and uses that as the return
type of the fetch-method.

As soon as we have C++14 function return type deduction, we can remove
that parameter again.