ObjectWrapper: Compare views by comparing the obj
Now comparison of ObjectHandles and ObjectViews "simply" compares the
pointers of the object, instead of using std::owner_less. This allows to
sort an unwrapped list of objects by pointer address and wrap it into
ObjectViews afterwards, retaining the sorting.

The drawback of that method is that comparing ObjectViews now requires
to lock the weak_ptr, but therefore the behavior of the comparison is as
you would expect it.