ObjWrapper: ForeignPointer be implicitly NonOwning
Owning foreign pointer means that the object-data is fetched at
first access to a property. This contradicts the idea of prefetching
everything while it's safe to do so and only serve from the cache
thereafter.

Having non owning foreign pointers devolves the user the duty to
prefetch the pointed to object at a suitable point in time before
the foreign pointer is accessed the first time. The foreign pointer will
just return a nullptr if accessed before.