ObjectWrapper: Use RTTI at objectwrapper creation
This allows us to create the shadow copy using the dynamic type of
the object not the static type of the pointer at hand.

To do this, for each type we create a static vector of subtype-factories
called by ObjectWrapperPrivate::create<Class>(). Each tries a
dynamic_cast<Subtype*> on the object, if it succeeds calls
ObjectWrapperPrivate::create<Subtype>().